@@ -8,38 +8,32 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build:es5:ivy": "npm run build -- --ts-config ./tsconfig.es5ivy.app.json --prod",
-    "build:es2015:ivy": "npm run build -- --ts-config ./tsconfig.es2015ivy.app.json --prod",
     "test": "npm run test:jasmine && npm run test:jest",
-    "test:debug": "npm run test:jasmine:es2015:ivy -- --browsers=Chrome --watch",
-    "test:jasmine": "npm run test:jasmine:es5:ivy && npm run test:jasmine:es2015:ivy",
-    "test:jasmine:es5:ivy": "ng test --ts-config ./tsconfig.es5ivy.spec.json --progress=false",
-    "test:jasmine:es2015:ivy": "ng test --ts-config ./tsconfig.es2015ivy.spec.json --progress=false",
-    "test:jasmine:debug": "ng test -- --watch --browsers Chrome",
-    "test:jest": "npm run test:jest:es5:ivy && npm run test:jest:es2015:ivy",
-    "test:jest:es5:ivy": "jest -w 1 --config jest.es5ivy.js",
-    "test:jest:es2015:ivy": "jest -w 1 --config jest.es2015ivy.js",
-    "test:jest:debug": "jest -w 1 --watch"
+    "test:debug": "npm run test:jasmine:debug",
+    "test:jasmine": "ng test",
+    "test:jasmine:debug": "npm run test:jasmine -- --browsers=Chrome --watch",
+    "test:jest": "jest -w 1 --config jest.config.ts",
+    "test:jest:debug": "npm run test:jest -- --watch"
   },
   "peerDependencies": {
     "ng-mocks": "*"
   },
   "dependencies": {
-    "@angular/common": "14.2.12",
-    "@angular/compiler": "14.2.12",
-    "@angular/core": "14.2.12",
-    "@angular/forms": "14.2.12",
-    "@angular/platform-browser": "14.2.12",
-    "@angular/platform-browser-dynamic": "14.2.12",
-    "@angular/router": "14.2.12",
+    "@angular/common": "15.0.1",
+    "@angular/compiler": "15.0.1",
+    "@angular/core": "15.0.1",
+    "@angular/forms": "15.0.1",
+    "@angular/platform-browser": "15.0.1",
+    "@angular/platform-browser-dynamic": "15.0.1",
+    "@angular/router": "15.0.1",
     "rxjs": "7.5.7",
     "tslib": "2.4.1",
     "zone.js": "0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "14.2.10",
-    "@angular/cli": "14.2.10",
-    "@angular/compiler-cli": "14.2.12",
+    "@angular-devkit/build-angular": "15.0.1",
+    "@angular/cli": "15.0.1",
+    "@angular/compiler-cli": "15.0.1",
     "@types/jasmine": "4.3.0",
     "@types/jest": "28.1.8",
     "@types/node": "16.18.3",
@@ -51,7 +45,7 @@
     "karma-chrome-launcher": "3.1.1",
     "karma-jasmine": "5.1.0",
     "karma-jasmine-html-reporter": "2.0.0",
-    "ng-packagr": "14.2.2",
+    "ng-packagr": "15.0.1",
     "npm": "8.19.3",
     "puppeteer": "19.3.0",
     "ts-node": "10.9.1",
