@@ -1,13 +1,11 @@
-module.exports = {
+export default {
   preset: 'jest-preset-angular',
   setupFilesAfterEnv: ['<rootDir>/src/setup-jest.ts'],
-  testEnvironmentOptions: {
-    url: 'http://localhost',
-  },
+  testURL: 'http://localhost',
   testPathIgnorePatterns: ['<rootDir>/src/test.ts'],
   globals: {
     'ts-jest': {
-      tsconfig: 'tsconfig.es5ivy.spec.json',
+      tsconfig: 'tsconfig.json',
     },
   },
 };
